+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# vNext
+# v0.4.16
 - Do not emit for `declare class`
 - Fixes issues where legacy metadata (`design:*`) was output without generating a suitable import, leading to `TypeError` at runtime